pyproject.toml
[tool.poetry]
name = "llm-translation-evaluator"
version = "0.1.0"
description = "LLM Translation Evaluator"
authors = ["Vincent <vincetecson27@gmail.com>"]
readme = "README.md"
packages = [
    { include = "llm_translation_evaluator", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = "^2.28.0"
langchain = ">=0.0.193"
openai = "^1.61.0"
anthropic = "^0.45.2"
evaluate = "^0.4.0"
sentence-transformers = "^2.2.2"
numpy = "*"
pandas = "*"
langchain-anthropic="0.3.6"
langchain-openai="0.3.3"
python-dotenv = "^1.0.1"
rich = "^13.9.4"
nltk = "^3.9.1"
rouge = "^1.0.1"
google-generativeai = "^0.8.4"




[build-system]
requires = [
  "poetry-core>=2.0.0,<3.0.0",
  "wheel",
  "Cython>=0.29.24",
  "numpy==1.23.5"
]
build-backend = "poetry.core.masonry.api"
